+++ src/muncher/items.js
@@ -28,6 +28,7 @@ async function generateImportItems(items) {
     name: "",
     flags: {
       ddbimporter: {
+        compendium: true,
         dndbeyond: {
           effectAbilities: [],
           totalLevels: 0,
@@ -64,7 +65,7 @@ async function generateImportItems(items) {
     }
   };
   let itemSpells = []; // here we need to parse each available spell and build a mock spell parser
-  const inventory = getInventory(mockDDB, mockCharacter, itemSpells, true);
+  const inventory = getInventory(mockDDB, mockCharacter, itemSpells);
   const results = {
     items: inventory,
     itemSpellNames: itemSpells, // this needs to be a list of spells to find