+++ src/parser/inventory/index.js
@@ -395,7 +395,7 @@ function enrichFlags(data, item) {
   if (data.definition?.sources) item.flags.ddbimporter.dndbeyond['sources'] = data.definition.sources;
 }
 
-export default function getInventory(ddb, character, itemSpells, compendiumItem = false) {
+export default function getInventory(ddb, character, itemSpells) {
   let items = [];
   // first, check custom name, price or weight
   ddb.character.characterValues.forEach((cv) => {
@@ -430,6 +430,7 @@ export default function getInventory(ddb, character, itemSpells, compendiumItem 
       item.flags.ddbimporter.originalName = originalName;
       if (!item.effects) item.effects = [];
       const daeInstalled = utils.isModuleInstalledAndActive("dae");
+      const compendiumItem = character.flags.ddbimporter.compendium;
       const addEffects = (compendiumItem)
         ? game.settings.get("ddb-importer", "munching-policy-add-effects")
         : game.settings.get("ddb-importer", "character-update-policy-add-effects");